2016/requirements.txt
nose
pyyaml
pypaths
